+++ src/components/code-blocks/index.tsx
@@ -15,8 +15,8 @@ import _V2 from "./v2.mdx"
 import _V3 from "./v3.mdx"
 import _V4 from "./v4.mdx"
 import _V5 from "./v5.mdx"
-import _Query from "./query.mdx"
-import _Schema from "./schema.mdx"
+export { default as Query } from "./query.mdx"
+export { default as Schema } from "./schema.mdx"
 import _QueryHeroFriends from "./query.hero-friends.mdx"
 import _ResponseHeroFriends from "./response.hero-friends.mdx"
 import _PredictableResult from "./predictable-result.mdx"
@@ -35,8 +35,6 @@ export const Code2 = () => <_Code2 components={components} />
 export const Code3 = () => <_Code3 components={components} />
 export const Code4 = () => <_Code4 components={components} />
 
-export const Query = () => <_Query components={components} />
-export const Schema = () => <_Schema components={components} />
 export const CodeA = () => <_CodeA components={components} />
 export const CodeB = () => <_CodeB components={components} />
 export const CodeC = () => <_CodeC components={components} />
